+++ tests/unit/config/test_config_validator.py
@@ -5,6 +5,7 @@ import pytest
 import yaml
 
 import esmvalcore
+import esmvalcore.cmor.table
 from esmvalcore import __version__ as current_version
 from esmvalcore.config import CFG
 from esmvalcore.config._config_validators import (
@@ -228,10 +229,26 @@ def generate_validator_testcases(valid):
         {
             "validator": validate_projects,
             "success": (
-                ({"CMIP6": {}}, {"CMIP6": {}}),
+                (
+                    {"CMIP6": {}},
+                    {
+                        "CMIP6": {
+                            "cmor_table": {
+                                "type": "esmvalcore.cmor.table.NoInfo",
+                            },
+                        },
+                    },
+                ),
                 (
                     {"CMIP6": {"extra_facets": {}}},
-                    {"CMIP6": {"extra_facets": {}}},
+                    {
+                        "CMIP6": {
+                            "cmor_table": {
+                                "type": "esmvalcore.cmor.table.NoInfo",
+                            },
+                            "extra_facets": {},
+                        },
+                    },
                 ),
             ),
             "fail": (
@@ -311,13 +328,14 @@ def test_handle_deprecation(remove_version):
 def test_validate_config_developer_none():
     """Test ``validate_config_developer``."""
     path = validate_config_developer(None)
-    assert path == Path(esmvalcore.__file__).parent / "config-developer.yml"
+    assert path is None
 
 
-def test_validate_config_developer(tmp_path):
+def test_validate_config_developer(tmp_path, monkeypatch):
     """Test ``validate_config_developer``."""
+    monkeypatch.setattr(esmvalcore.cmor.table, "CMOR_TABLES", {})
     custom_table_path = (
-        Path(esmvalcore.__file__).parent / "cmor" / "tables" / "custom"
+        Path(esmvalcore.__file__).parent / "cmor" / "tables" / "cmip5-custom"
     )
     cfg_dev = {
         "custom": {"cmor_path": custom_table_path},